room이 정의되어 있지 않던 문제 해결

+++ nodejs/Services/together/togetherInviteService.js
@@ -18,10 +18,15 @@ const togetherInviteService = (socket, msg) => {
 
     // 기존 접속 방에서 나가기
     if (user.joinedRoomName !== null) {
-      if (room.deleteUser(user) === 0) {
-        SingleObject.RoomRepository.delete(room.getRoomName());
-      } else {
-        gameRoomUpdateService(socket, roomName, id);
+      const room = SingleObject.RoomRepository.findByRoomName(
+        user.joinedRoomName
+      );
+      if (room !== false) {
+        if (room.deleteUser(user) === 0) {
+          SingleObject.RoomRepository.delete(room.getRoomName());
+        } else {
+          gameRoomUpdateService(socket, roomName, id);
+        }
       }
       user.updateJoinedRoomName(null);
     }
